requirements.txt
sounddevice
numpy
scipy
pystray
Pillow
pynput
pyperclip
pydantic
